refactor: 토큰 비즈니스로직 수정 - 2
- 불필요한 클래스 제거: 대기열 관리 클래스

+++ src/main/java/io/hhplus/conbook/domain/token/TokenStatusCount.java
@@ -8,9 +8,9 @@ import lombok.ToString;
 @ToString
 public class TokenStatusCount {
     private TokenStatus status;
-    private int count;
+    private long count;
 
-    public TokenStatusCount(TokenStatus status, int count) {
+    public TokenStatusCount(TokenStatus status, long count) {
         this.status = status;
         this.count = count;
     }
